(#94)refactor:  constants 위치 수정했습니다.
