+++ src/components/pages/storePage/storePage.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import '../../../asset/sass/pages/storePage/storePage.scss';
-import { ACCESS_TOKEN } from '../../pages/loginPage/constants/index.js';
+import { ACCESS_TOKEN } from '../../global/constants/index.js';
 import { StyledPage, StyledHeader } from '../../../styledComponent.js';
 import TitleHeader from '../../ui/header/titleHeader.jsx';
 import TabBar from '../../ui/tabBar/tabBar.jsx';
