+++ src/components/ui/header/userInfoHeader.jsx
@@ -13,7 +13,7 @@ import {
   BASE_URL,
   ACCESS_TOKEN,
   REFRESH_TOKEN,
-} from '../../pages/loginPage/constants';
+} from '../../global/constants/index';
 
 function UserInfoHeader() {
   const { isLoggedIn, rewardCount, isDropdownOpen } = useSelector(
