+++ src/router/allRouter.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { Route, Routes } from 'react-router-dom';
-import { ACCESS_TOKEN } from '../components/pages/loginPage/constants/index.js';
+import { ACCESS_TOKEN } from '../components/global/constants/index.js';
 
 /* 메인 페이지 관련 impot */
 import MainPage from '../components/pages/mainPage/mainPage';
